Fixing issue #28
The issue was actually quite simple; the equality comparer for type
references was verifying that actual subtypes were the same, such that a
TypeReference to Foo was not equal to the TypeDefinition of Foo.  Removing
that check resolves this issue nicely.

There may be implications further down the road, as now any two
TypeReferences with the same FullName are equal, but I think that's what I
intended anyways.  This could potentially allow for version-mismatching in
references, but that's a problem which has yet to surface.